+++ idea-plugin/src/main/kotlin/com/dmxiaoshubao/copypathline/idea/CopyPathLineBundle.kt
@@ -1,8 +1,0 @@
-package com.dmxiaoshubao.copypathline.idea
-
-import com.intellij.AbstractBundle
-import org.jetbrains.annotations.PropertyKey
-
-object CopyPathLineBundle : AbstractBundle("messages.CopyPathLineBundle") {
-    fun message(@PropertyKey(resourceBundle = "messages.CopyPathLineBundle") key: String): String = getMessage(key)
-}