+++ idea-plugin/src/main/kotlin/com/dmxiaoshubao/copypathline/idea/actions/CopyRelativePathWithLineAction.kt
@@ -1,7 +1,0 @@
-package com.dmxiaoshubao.copypathline.idea.actions
-
-import com.dmxiaoshubao.copypathline.idea.core.CopyMode
-
-class CopyRelativePathWithLineAction : BaseCopyPathWithLineAction() {
-    override fun copyMode(): CopyMode = CopyMode.RELATIVE
-}