+++ idea-plugin/src/main/kotlin/com/dmxiaoshubao/copypathline/idea/core/CopyMode.kt
@@ -1,6 +1,0 @@
-package com.dmxiaoshubao.copypathline.idea.core
-
-enum class CopyMode {
-    ABSOLUTE,
-    RELATIVE,
-}