+++ idea-plugin/src/main/kotlin/com/dmxiaoshubao/copypathline/idea/core/LineRange.kt
@@ -1,3 +1,0 @@
-package com.dmxiaoshubao.copypathline.idea.core
-
-data class LineRange(val start: Int, val end: Int)