+++ idea-plugin/src/main/kotlin/com/dmxiaoshubao/copypathline/idea/core/SelectionRange.kt
@@ -1,7 +1,0 @@
-package com.dmxiaoshubao.copypathline.idea.core
-
-data class SelectionRange(
-    val startOffset: Int,
-    val endOffset: Int,
-    val hasSelection: Boolean,
-)